+++ jib-gradle-plugin/src/main/java/com/google/cloud/tools/jib/gradle/GradleRawConfiguration.java
@@ -132,11 +132,6 @@ class GradleRawConfiguration implements RawConfiguration {
   }
 
   @Override
-  public boolean getUseCurrentTimestamp() {
-    return jibExtension.getContainer().getUseCurrentTimestamp();
-  }
-
-  @Override
   public boolean getAllowInsecureRegistries() {
     return jibExtension.getAllowInsecureRegistries();
   }