+++ jib-plugins-common/src/main/java/com/google/cloud/tools/jib/plugins/common/PluginConfigurationProcessor.java
@@ -217,15 +217,8 @@ public class PluginConfigurationProcessor {
             .setUser(rawConfiguration.getUser().orElse(null));
     getWorkingDirectoryChecked(rawConfiguration)
         .ifPresent(jibContainerBuilder::setWorkingDirectory);
-    if (rawConfiguration.getUseCurrentTimestamp()) {
-      projectProperties.log(
-          LogEvent.warn(
-              "Setting image creation time to current time; your image may not be reproducible."));
-      jibContainerBuilder.setCreationTime(Instant.now());
-    } else {
-      jibContainerBuilder.setCreationTime(
-          getCreationTime(rawConfiguration.getCreationTime(), projectProperties));
-    }
+    jibContainerBuilder.setCreationTime(
+        getCreationTime(rawConfiguration.getCreationTime(), projectProperties));
 
     // Adds all the extra files.
     for (Path directory : rawConfiguration.getExtraDirectories()) {