+++ jib-plugins-common/src/main/java/com/google/cloud/tools/jib/plugins/common/PropertyNames.java
@@ -62,8 +62,5 @@ public class PropertyNames {
   public static final String CONTAINERIZE = "jib.containerize";
   public static final String ALWAYS_CACHE_BASE_IMAGE = "jib.alwaysCacheBaseImage";
 
-  @Deprecated
-  public static final String CONTAINER_USE_CURRENT_TIMESTAMP = "jib.container.useCurrentTimestamp";
-
   private PropertyNames() {}
 }